Renamed js_user_apikey_provider.hpp to js_api_key_auth.hpp
